App element relocated and renamed
